spot/requirements.txt
bosdyn-client
bosdyn-mission
bosdyn-choreography-client
scipy
robonomics-interface==1.0.5
substrate-interface>=1.2.5,<2
requests
Flask
Flask-Cors
opencv-python
python-dotenv
git+https://github.com/Smehnov/pinatapy
